requirements.txt
kafka-python==2.0.2
faker==18.13.0
docker-compose==1.29.2
